refactor(migrations): 清理28个冗余迁移脚本，优化迁移目录结构
主要变更:
- 删除28个已被UNIFIED_MIGRATION_v1.sql合并的冗余迁移脚本
- 迁移文件移至docker/migrations/目录备份
- 更新MIGRATION_INDEX_V3.md至v4.0.0
- 添加新迁移: device_trust_tables, secure_backup_tables

迁移目录结构优化:
- 核心schema: 00000000_unified_schema_v6.sql
- 统一迁移: UNIFIED_MIGRATION_v1.sql (合并28个增量迁移)
- 增量迁移: 6个字段重命名和功能增强脚本

+++ src/web/routes/key_rotation.rs
@@ -120,7 +120,7 @@ pub async fn revoke_old_keys(
         .await;
 
         if let Ok(r) = result {
-            revoked_count += r.rows_affected() as u64;
+            revoked_count += r.rows_affected();
         }
     }
